LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_1-script-src-expected.txt

default-src should cascade to script-src directive


PASS Verify cascading of default-src to script-src policy: block
PASS Verify cascading of default-src to script-src policy: allow
FAIL Should fire violation events for every failed violation assert_equals: expected "script-src-elem" but got "default-src"
